requirements.txt
# Core dependencies - with compatible versions
tensorflow==2.13.0
tensorflow-federated==0.50.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
joblib==1.3.1
click==8.1.6

# Privacy alternatives (compatible versions)
# diffprivlib - comment out if causing issues
# diffprivlib==0.6.1  

# PySyft - remove as it's not essential and has compatibility issues
# pysyft==0.7.0

# Optional privacy library (if you need differential privacy)
opacus==1.4.0  # Facebook's privacy library, more stable

# Notebooks
jupyter==1.0.0
ipykernel==6.25.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Reporting
jinja2==3.1.2
# pdfkit==1.0.0  # Comment out if you don't need PDF generation
